requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=1.5.0
json5>=0.9.0
fastapi>=0.111.0
uvicorn>=0.30.0